- Move `ShelleyEra` and associated definitions to `Cardano.Ledger.Shelley`. - Define `Core` instances in the `ShelleMA` era module.
