+++ shelley-ma/impl/cardano-ledger-shelley-ma.cabal
@@ -29,7 +29,6 @@ library
     Cardano.Ledger.ShelleyMA.Metadata
     Cardano.Ledger.ShelleyMA.Rules.Utxo
     Cardano.Ledger.ShelleyMA.Rules.Utxow
-    Cardano.Ledger.ShelleyMA.Scripts
     Cardano.Ledger.ShelleyMA.Timelocks
     Cardano.Ledger.ShelleyMA.TxBody
