+++ shelley-ma/impl/src/Cardano/Ledger/Allegra.hs
@@ -6,10 +6,7 @@
 module Cardano.Ledger.Allegra where
 
 import Cardano.Ledger.ShelleyMA
-import Cardano.Ledger.ShelleyMA.Metadata ()
-import Cardano.Ledger.ShelleyMA.Rules.Utxo ()
 import Cardano.Ledger.ShelleyMA.Rules.Utxow ()
-import Cardano.Ledger.ShelleyMA.Scripts ()
 import Cardano.Ledger.ShelleyMA.TxBody ()
 import Shelley.Spec.Ledger.API (ApplyBlock, ApplyTx, GetLedgerView, PraosCrypto, ShelleyBasedEra)
