+++ shelley/chain-and-ledger/executable-spec/src/Shelley/Spec/Ledger/Tx.hs
@@ -72,10 +72,8 @@ import Cardano.Binary
     withSlice,
   )
 import qualified Cardano.Ledger.Core as Core
-import qualified Cardano.Ledger.Crypto as CryptoClass
 import Cardano.Ledger.Era
-import Cardano.Ledger.Shelley (ShelleyBased, ShelleyEra, TxBodyConstraints)
-import qualified Cardano.Ledger.Shelley as Shelley
+import Cardano.Ledger.Shelley.Constraints (ShelleyBased, TxBodyConstraints)
 import qualified Data.ByteString.Lazy as BSL
 import Data.Foldable (fold)
 import Data.Functor.Identity (Identity)
@@ -218,7 +216,7 @@ deriving instance
   Eq (Tx era)
 
 pattern Tx ::
-  ( Shelley.TxBodyConstraints era,
+  ( TxBodyConstraints era,
     ToCBOR (Core.Metadata era)
   ) =>
   Core.TxBody era ->
@@ -251,7 +249,7 @@ instance ShelleyBased era => HashAnnotated (Tx era) era where
   type HashIndex (Tx era) = EraIndependentTx
 
 segwitTx ::
-  ( Shelley.TxBodyConstraints era,
+  ( TxBodyConstraints era,
     ToCBOR (Core.Metadata era)
   ) =>
   Annotator (Core.TxBody era) ->
@@ -282,7 +280,7 @@ segwitTx
 
 decodeWits ::
   forall era s.
-  ( Shelley.TxBodyConstraints era,
+  ( TxBodyConstraints era,
     Core.AnnotatedData (Core.Script era),
     ValidateScript era
   ) =>
@@ -357,14 +355,6 @@ class
   validateScript :: Core.Script era -> Tx era -> Bool
   hashScript :: Core.Script era -> ScriptHash era
 
--- | instance of MultiSignatureScript type class
-instance
-  (CryptoClass.Crypto c, TxBodyConstraints (ShelleyEra c)) =>
-  ValidateScript (ShelleyEra c)
-  where
-  validateScript = validateNativeMultiSigScript
-  hashScript = hashMultiSigScript
-
 -- | Script evaluator for native multi-signature scheme. 'vhks' is the set of
 -- key hashes that signed the transaction to be validated.
 evalNativeMultiSigScript ::
@@ -382,7 +372,7 @@ evalNativeMultiSigScript (RequireMOf m msigs) vhks =
 
 -- | Script validator for native multi-signature scheme.
 validateNativeMultiSigScript ::
-  ( Shelley.TxBodyConstraints era,
+  ( TxBodyConstraints era,
     ToCBOR (Core.Metadata era)
   ) =>
   MultiSig era ->
@@ -396,7 +386,7 @@ validateNativeMultiSigScript msig tx =
 
 -- | Multi-signature script witness accessor function for Transactions
 txwitsScript ::
-  (Shelley.TxBodyConstraints era, ToCBOR (Core.Metadata era)) =>
+  (TxBodyConstraints era, ToCBOR (Core.Metadata era)) =>
   Tx era ->
   Map (ScriptHash era) (Core.Script era)
 txwitsScript = scriptWits' . _witnessSet
